requirements.txt
Pillow
django-filer